chore: better AST file structure
